Extend error wrapping to global funcs
